+++ src/archive.cpp
@@ -284,7 +284,7 @@ namespace librealsense
                 auto callback_duration = callback_ended - frame->get_frame_callback_start_time_point();
 
                 LOG_DEBUG("CallbackFinished," << librealsense::get_string(frame->get_stream()->get_stream_type()) << "," << frame->get_frame_number()
-                    << ",DispatchedAt," << callback_ended);
+                    << ",DispatchedAt," << std::fixed << callback_ended);
 
                 if (callback_duration > callback_warning_duration)
                 {
@@ -498,7 +498,7 @@ namespace librealsense
     void frame::log_callback_start(rs2_time_t timestamp)
     {
         update_frame_callback_start_ts(timestamp);
-        LOG_DEBUG("CallbackStarted," << std::dec << librealsense::get_string(get_stream()->get_stream_type()) << "," << get_frame_number() << ",DispatchedAt," << timestamp);
+        LOG_DEBUG("CallbackStarted," << std::dec << librealsense::get_string(get_stream()->get_stream_type()) << "," << get_frame_number() << ",DispatchedAt," << std::fixed << timestamp);
     }
 
     void frame::log_callback_end(rs2_time_t timestamp) const
@@ -506,7 +506,7 @@ namespace librealsense
         auto callback_warning_duration = 1000.f / (get_stream()->get_framerate() + 1);
         auto callback_duration = timestamp - get_frame_callback_start_time_point();
 
-        LOG_DEBUG("CallbackFinished," << librealsense::get_string(get_stream()->get_stream_type()) << "," << get_frame_number() << ",DispatchedAt," << timestamp);
+        LOG_DEBUG("CallbackFinished," << librealsense::get_string(get_stream()->get_stream_type()) << "," << get_frame_number() << ",DispatchedAt," << std::fixed << timestamp);
 
         if (callback_duration > callback_warning_duration)
         {
