+++ src/linux/backend-v4l2.h
@@ -204,7 +204,7 @@ namespace librealsense
                 {
                     if (_data_buf && (!_managed))
                     {
-                        LOG_DEBUG("Enqueue buf " << _dq_buf.index << " for fd " << _file_desc);
+                        //LOG_DEBUG("Enqueue buf " << _dq_buf.index << " for fd " << _file_desc);
                         if (xioctl(_file_desc, (int)VIDIOC_QBUF, &_dq_buf) < 0)
                         {
                             LOG_ERROR("xioctl(VIDIOC_QBUF) guard failed");
