+++ tools/data-collect/rs-data-collect.cpp
@@ -1,311 +1,19 @@
 // License: Apache 2.0. See LICENSE file in root directory.
-// Copyright(c) 2015 Intel Corporation. All Rights Reserved.
-// The utility shall be used to collect and analyze Realsense Cameras performance.
-// Therefore no UI is provided, and the data is gathered and serialized in csv-compatible format for offline analysis.
-// The utility is configured via command-line parameters and requires user-provided configuration file to run
-// The lines 8-20 constitute a valid configuration file for D435i camera
-/*
-//Defines the order at which the params (comma-separated) should appear in configuration file
-//#rs2_stream, width, height, fps, rs2_format, stream_index (optional, put zero or leave empty if you're not sure)
-//# Examples (note that all lines starting with non-alpha character (#@!...) will be skipped:
-//  DEPTH,640,480,30,Z16,0
-//  INFRARED,640,480,30,Y8,1
-//  COLOR,640,480,30,RGB8,0
-//  ACCEL,1,1,125,MOTION_XYZ32F
-//  GYRO,1,1,200,MOTION_XYZ32F
-DEPTH,640,480,30,Z16,0
-INFRARED,640,480,30,Y8,1
-COLOR,640,480,30,RGB8,0
-ACCEL,1,1,125,MOTION_XYZ32F
-GYRO,1,1,200,MOTION_XYZ32F
-*/
+// Copyright(c) 2018 Intel Corporation. All Rights Reserved.
+// Minimalistic command-line collect & analyze bandwidth/performance tool for Realsense Cameras.
+// The data is gathered and serialized in csv-compatible format for offline analysis.
+// Extract and store frame headers info for video streams; for IMU&Tracking streams also store the actual data
+// The utility is configured with command-line keys and requires user-provided config file to run
+// See rs-data-collect.h for config examples
 
 #include <librealsense2/rs.hpp>
-#include <chrono>
+#include "rs-data-collect.h"
+#include "tclap/CmdLine.h"
 #include <thread>
 #include <iostream>
-#include <fstream>
-#include <sstream>
-#include "tclap/CmdLine.h"
-#include <condition_variable>
-#include <set>
-#include <cctype>
-#include <thread>
-#include <array>
-#include <map>
-#include <atomic>
-#include <regex>
-#include <algorithm>    // std::sort
 
 using namespace std;
 using namespace TCLAP;
-
-const uint64_t  DEF_FRAMES_NUMBER = 100;
-const std::string DEF_OUTPUT_FILE_NAME("frames_data.csv");
-
-// Split string into token,  trim unreadable characters
-std::vector<std::string> tokenize(std::string line, char separator)
-{
-    std::vector<std::string> tokens;
-
-    // Remove trailing characters
-    line.erase(line.find_last_not_of("\t\n\v\r ") + 1);
-
-    stringstream ss(line);
-    while (ss.good())
-    {
-        string substr;
-        getline(ss, substr, separator);
-        tokens.push_back(substr);
-    }
-
-    return tokens;
-}
-
-struct stringify
-{
-    std::ostringstream ss;
-    template<class T> stringify & operator << (const T & val) { ss << val; return *this; }
-    operator std::string() const { return ss.str(); }
-};
-
-struct frame_data
-{
-    frame_data(unsigned long long frame_number,
-               double frame_ts,
-               double host_ts,
-               rs2_timestamp_domain domain,
-               rs2_stream stream_type,
-               int stream_index):
-    _frame_number(frame_number),
-    _ts(frame_ts),
-    _arrival_time(host_ts),
-    _domain(domain),
-    _stream_type(stream_type),
-    _stream_idx(stream_index)
-    {};
-
-    virtual std::string to_string() const
-    {
-        std::stringstream ss;
-        ss  << std::endl
-            <<  rs2_stream_to_string(_stream_type) << ","
-            << _stream_idx << "," << _frame_number << ","
-            << std::fixed << std::setprecision(3) << _ts << "," << _arrival_time;
-
-        return ss.str().c_str();
-    }
-
-    unsigned long long      _frame_number;
-    double                  _ts;
-    double                  _arrival_time;
-    rs2_timestamp_domain    _domain;
-    rs2_stream              _stream_type;
-    int                     _stream_idx;
-};
-
-struct threedof_data : frame_data
-{
-    threedof_data(unsigned long long frame_number,
-               double frame_ts,
-               double host_ts,
-               rs2_timestamp_domain domain,
-               rs2_stream stream_type,
-               int stream_index,
-               double x, double y, double z):
-        frame_data(frame_number, frame_ts, host_ts, domain, stream_type, stream_index),
-        _x_axis(x), _y_axis(y), _z_axis(z)
-        {};
-
-    virtual std::string to_string() const override
-    {
-        std::stringstream ss;
-        ss  << frame_data::to_string()<< ","
-            << _x_axis << "," << _y_axis << "," << _z_axis;
-
-        return ss.str().c_str();
-    }
-
-    double  _x_axis;
-    double  _y_axis;
-    double  _z_axis;
-};
-
-struct stream_request
-{
-    rs2_stream  _stream_type;
-    rs2_format  _stream_format;
-    int         _width;
-    int         _height;
-    int         _fps;
-    int         _stream_idx;
-};
-
-inline std::ostream&  operator<<(std::ostream& os, const stream_request& req)
-{
-    return os << "Type: " << req._stream_type << ",Idx: " << req._stream_idx << ", "
-        << req._stream_format << ", [" << req._width << "x" << req._height << "], " << req._fps << "fps" << std::endl;
-}
-
-enum config_params { STREAM_TYPE = 0, RES_WIDTH, RES_HEIGHT, FPS, FORMAT, STREAM_INDEX };
-
-int parse_number(char const *s, int base = 0)
-{
-    char c;
-    stringstream ss(s);
-    int i;
-    ss >> i;
-
-    if (ss.fail() || ss.get(c))
-    {
-        throw runtime_error(string(string("Invalid numeric input - ") + s + string("\n")));
-    }
-    return i;
-}
-
-std::string to_lower(const std::string& s)
-{
-    auto copy = s;
-    std::transform(copy.begin(), copy.end(), copy.begin(), ::tolower);
-    return copy;
-}
-
-rs2_format parse_format(const string str)
-{
-    for (int i = RS2_FORMAT_ANY; i < RS2_FORMAT_COUNT; i++)
-    {
-        if (to_lower(rs2_format_to_string((rs2_format)i)) == str)
-        {
-            return (rs2_format)i;
-        }
-    }
-    throw runtime_error((string("Invalid format - ") + str + string("\n")).c_str());
-}
-
-rs2_stream parse_stream_type(const string str)
-{
-    for (int i = RS2_STREAM_ANY; i < RS2_STREAM_COUNT; i++)
-    {
-        if (to_lower(rs2_stream_to_string((rs2_stream)i)) == str)
-        {
-            return static_cast<rs2_stream>(i);
-        }
-    }
-    throw runtime_error((string("Invalid stream type - ") + str + string("\n")).c_str());
-}
-
-int parse_fps(const string str)
-{
-    return parse_number(str.c_str());
-}
-
-bool parse_configuration(const std::string& line, const std::vector<std::string>& tokens,
-    rs2_stream& type, int& width, int& height, rs2_format& format, int& fps, int& index)
-{
-    bool res = false;
-    try
-    {
-        auto tokens = tokenize(line,',');
-        if (tokens.size() < STREAM_INDEX)
-            return res;
-
-        // Convert string to uppercase
-        type    =   parse_stream_type(to_lower(tokens[STREAM_TYPE]));
-        width   =   parse_number(tokens[RES_WIDTH].c_str());
-        height  =   parse_number(tokens[RES_HEIGHT].c_str());
-        fps     =   parse_fps(tokens[FPS]);
-        format  =   parse_format(to_lower(tokens[FORMAT]));
-        // Backward compatibility
-        if (tokens.size() > STREAM_INDEX)
-            index = parse_number(tokens[STREAM_INDEX].c_str());
-        res = true;
-        std::cout << "Request added for " << line << std::endl;
-    }
-    catch (...)
-    {
-        std::cout << "Invalid syntax in configuration line " << line << std::endl;
-    }
-
-    return res;
-}
-
-void parse_requests(std::vector<stream_request>& reqs, const std::string& config_filename)
-{
-    ifstream file(config_filename);
-
-    if (!file.is_open())
-        throw runtime_error(stringify() << "Given .csv configure file " << config_filename <<  " was not Found!");
-
-    // Line starting with non-alpha characters will be treated as comments
-    const std::regex starts_with("^[a-zA-Z]");
-    string line;
-    rs2_stream stream_type;
-    rs2_format format;
-    int width{}, height{}, fps{}, index{};
-
-    while (getline(file, line))
-    {
-        // Parsing configuration requests
-        auto tokens =tokenize(line,',');
-
-        if (std::regex_search(line, starts_with))
-        {
-            if (parse_configuration(line, tokens, stream_type, width, height, format, fps,index))
-                reqs.push_back({ stream_type, format, width, height, fps,  index });
-        }
-    }
-
-    // Sanity test agains multiple conflicting requests for single sensor
-    if (reqs.size())
-    {
-        std::sort(reqs.begin(), reqs.end(), [](const stream_request& l, const stream_request& r) { return l._stream_type < r._stream_type; });
-        for (auto i = 0; i < reqs.size() - 1; i++)
-        {
-            if ((reqs[i]._stream_type == reqs[i+1]._stream_type) && ((reqs[i]._stream_idx == reqs[i+1]._stream_idx)))
-                throw runtime_error(stringify() << "Invalid configuration file - multiple requests for the same sensor:\n"
-                    << reqs[i] << "\n" << reqs[+i]);
-        }
-    }
-    else
-        throw std::runtime_error("Invalid configuration file - zero requests accepted");
-}
-
-std::string get_profile_description(const rs2::stream_profile& profile)
-{
-    std::stringstream ss;
-
-    ss  << profile.stream_name() << ","
-        << profile.stream_type() << ","
-        << profile.format() << ","
-        << profile.fps();
-
-    if (auto vp = profile.as<rs2::video_stream_profile>())
-        ss << "," << vp.width() << "," << vp.height();
-
-    ss << std::endl;
-    return ss.str().c_str();
-}
-
-void save_data_to_file(const std::map<std::pair<rs2_stream, int>, std::vector<std::shared_ptr<frame_data>>>& buffer,
-                        const std::vector<rs2::stream_profile>& active_stream_profiles,
-                        const string& out_filename)
-{
-    // Save/override to file
-    ofstream csv(out_filename);
-    //csv.open(out_filename);
-
-    csv << "Configuration:\nStream Type,Stream Name,Format,FPS,Width,Height\n";
-    for (const auto& elem : active_stream_profiles)
-        csv << get_profile_description(elem);
-
-    csv << "\nStream Type,Index,F#,HW Timestamp (ms),Host Timestamp(ms), 3DOF_x,3DOF_y,3DOF_z\n";
-
-    for (const auto& elem : buffer)
-    {
-        for (auto i = 0; i < elem.second.size(); i++)
-            csv << elem.second[i]->to_string();
-    }
-}
 
 enum application_stop : uint8_t {
     stop_on_frame_num,
@@ -314,6 +22,7 @@ enum application_stop : uint8_t {
     stop_on_any
 };
 
+using namespace rs_data_collect;
 
 int main(int argc, char** argv) try
 {
@@ -332,7 +41,7 @@ int main(int argc, char** argv) try
     cmd.add(config_file);
     cmd.parse(argc, argv);
 
-    std::cout << "Running rs-data-collect ";
+    std::cout << "Running rs-data-collect: ";
     for (auto i=1; i < argc; ++i)
         std::cout << argv[i] << " ";
     std::cout << std::endl;
@@ -350,67 +59,64 @@ int main(int argc, char** argv) try
 
     while (!succeed)
     {
-        std::vector<stream_request> requests, orig_requests;
-        
+        std::vector<stream_request> requests_to_go, user_requests;
         list = ctx.query_devices();
 
         if (0== list.size())
+        {
+            std::cout << "Connect Realsense Camera to proceed" << std::endl;
+            std::this_thread::sleep_for(std::chrono::seconds(3));
             continue;
+        }
 
         auto dev = std::make_shared<rs2::device>(list.front());
 
         if (config_file.isSet())
         {
-            parse_requests(requests, config_file.getValue());
+            parse_requests(user_requests, config_file.getValue());
         }
 
-        orig_requests = requests;
+        if (!config_file.isSet() || !user_requests.size())
+            throw std::runtime_error(stringify()
+                    << "The tool requires a profile configuration file to proceed"
+                    << "\nRun rs-data-collect --help for details");
+
+        requests_to_go = user_requests;
         std::vector<rs2::stream_profile> matches, total_matches;
+
         // Configure and starts streaming
         for (auto&& sensor : dev->query_sensors())
         {
             for (auto& profile : sensor.get_stream_profiles())
             {
                 // All requests have been resolved
-                if (!requests.size())
+                if (!requests_to_go.size())
                     break;
 
                 // Find profile matches
-                auto fulfilled_request = std::find_if(requests.begin(), requests.end(), [&matches, profile](const stream_request& req)
+                auto fulfilled_request = std::find_if(requests_to_go.begin(), requests_to_go.end(), [&matches, profile](const stream_request& req)
                 {
                     bool res = false;
-                    if (auto vp = profile.as<rs2::video_stream_profile>())
+                    if ((profile.stream_type()  == req._stream_type) &&
+                        (profile.format()       == req._stream_format) &&
+                        (profile.stream_index() == req._stream_idx) &&
+                        (profile.fps()          == req._fps))
                     {
-                        if ((vp.stream_type() == req._stream_type) &&
-                            (vp.format() == req._stream_format) &&
-                            (vp.stream_index() == req._stream_idx) &&
-                            (vp.fps() == req._fps) &&
-                            (vp.width() == req._width) &&
-                            (vp.height() == req._height))
+                        if (auto vp = profile.as<rs2::video_stream_profile>())
                         {
-                            matches.push_back(profile);
-                            res = true;
+                            if ((vp.width() != req._width) || (vp.height() != req._height))
+                                return false;
                         }
+                        res = true;
+                        matches.push_back(profile);
                     }
-                    else
-                    {
-                        if (auto mp = profile.as<rs2::motion_stream_profile>())
-                        {
-                            if ((mp.stream_type() == req._stream_type) &&
-                                (mp.format() == req._stream_format) &&
-                                (mp.stream_index() == req._stream_idx) &&
-                                (mp.fps() == req._fps))
-                            {
-                                matches.push_back(profile);
-                                res = true;
-                            }
-                        }
-                    }
+
                     return res;
                 });
 
-                if (fulfilled_request != requests.end())
-                    requests.erase(fulfilled_request);
+                // Remove the request once resolved
+                if (fulfilled_request != requests_to_go.end())
+                    requests_to_go.erase(fulfilled_request);
             }
 
             // Aggregate resolved requests
@@ -422,15 +128,29 @@ int main(int argc, char** argv) try
                 matches.clear();
             }
 
-            if (total_matches.size() == orig_requests.size())
+            if (total_matches.size() == user_requests.size())
                 succeed = true;
         }
 
-        std::cout << "\nDevice selected: " << dev->get_info(RS2_CAMERA_INFO_NAME)
-                << " s.n: " << dev->get_info(RS2_CAMERA_INFO_SERIAL_NUMBER)
-                << "\nMatching profiles: " << total_matches.size() << std::endl << std::endl;
+        std::cout << "\nDevice selected: \n\t" << dev->get_info(RS2_CAMERA_INFO_NAME)
+                << (dev->supports(RS2_CAMERA_INFO_USB_TYPE_DESCRIPTOR) ?
+                    std::string(( stringify() << ". USB Type: " << dev->get_info(RS2_CAMERA_INFO_USB_TYPE_DESCRIPTOR))) : "")
+                << "\n\tS.N: " << (dev->supports(RS2_CAMERA_INFO_SERIAL_NUMBER) ? dev->get_info(RS2_CAMERA_INFO_SERIAL_NUMBER) : "")
+                << "\n\tFW Ver: " << dev->get_info(RS2_CAMERA_INFO_FIRMWARE_VERSION)
+                << "\nUser streams requested: " << user_requests.size()
+                << ", actual matches: " << total_matches.size() << std::endl;
 
-        std::map<std::pair<rs2_stream, int>, std::vector<std::shared_ptr<frame_data>>> buffer;
+        if (!succeed)
+        {
+            std::cout << "\nThe following request(s) are not supported by the device: " << std::endl;
+            for(auto& elem : requests_to_go)
+                std::cout << elem;
+
+            if (requests_to_go.size() == user_requests.size())
+                throw std::runtime_error("\nNo valid stream requests were found, aborting..");
+        }
+
+        data_collection buffer;
         auto start_time = chrono::high_resolution_clock::now();
 
         // Start streaming
@@ -446,14 +166,6 @@ int main(int argc, char** argv) try
 
                 if (buffer[stream_uid].size() < max_frames_number)
                 {
-//                    threedof_data data{ f.get_frame_number(),
-//                        f.get_timestamp(),
-//                        arrival_time.count(),
-//                        f.get_frame_timestamp_domain(),
-//                        f.get_profile().stream_type(),
-//                        f.get_profile().stream_index(),
-//                        axes.x, axes.y, axes.z};
-//                    //buffer[stream_uid].push_back(data);
                     buffer[stream_uid].push_back(std::make_shared<threedof_data>(f.get_frame_number(),
                         f.get_timestamp(),
                         arrival_time.count(),
@@ -465,19 +177,37 @@ int main(int argc, char** argv) try
             }
             else
             {
-                if (buffer[stream_uid].size() < max_frames_number)
+                if (auto pf = f.as<rs2::pose_frame>())
                 {
-                    buffer[stream_uid].push_back(std::make_shared<frame_data>( f.get_frame_number(),
-                            f.get_timestamp(),
+                    if (buffer[stream_uid].size() < max_frames_number)
+                    {
+                        auto pose = pf.get_pose_data();
+                        buffer[stream_uid].push_back(std::make_shared<sixdof_data>(pf.get_frame_number(),
+                            pf.get_timestamp(),
                             arrival_time.count(),
                             f.get_frame_timestamp_domain(),
                             f.get_profile().stream_type(),
-                            f.get_profile().stream_index()));
+                            f.get_profile().stream_index(),
+                            pose.translation.x, pose.translation.y, pose.translation.z,
+                            pose.rotation.x,pose.rotation.y,pose.rotation.z,pose.rotation.w));
+                    }
+                }
+                else
+                {
+                    if (buffer[stream_uid].size() < max_frames_number)
+                    {
+                        buffer[stream_uid].push_back(std::make_shared<frame_data>( f.get_frame_number(),
+                                f.get_timestamp(),
+                                arrival_time.count(),
+                                f.get_frame_timestamp_domain(),
+                                f.get_profile().stream_type(),
+                                f.get_profile().stream_index()));
+                    }
                 }
             }
         });
 
-        std::cout << "Data collect started.... " << std::endl;
+        std::cout << "\nData collection started.... " << std::endl;
         bool timed_out = false;
 
         const auto collection_accomplished = [&]()
@@ -519,14 +249,11 @@ int main(int argc, char** argv) try
                     << " sec" << std::endl;
         }
 
-        try
+        // Stop & flush all active sensors
+        for (auto&& sensor : active_sensors)
         {
-            // Stop sensors streaming
-            for (auto&& sensor : dev->query_sensors())
-                sensor.stop();
-        }
-        catch(...){
-            std::cout <<" sensor.stop(); failed" << std::endl;
+            sensor.stop();
+            sensor.close();
         }
 
         std::vector<uint64_t> frames_per_stream;
@@ -543,7 +270,7 @@ int main(int argc, char** argv) try
 
         succeed = true;
     }
-    std::cout << "Job completed" << std::endl;
+    std::cout << "Task completed" << std::endl;
 
     return EXIT_SUCCESS;
 }
